Fixing horizontal rule error.
The error caused by "---- ----" was coming up because the rule was trying to match text
after the first "----".  The text rule then checked to avoid eating special lines, like
"----" or "* bullet".  The check was hitting on the second "----", causing the
text to not match, which meant nothing matched.

The fix is to remove the initial check in the text rule by creating unsafe_text
that matches a phrase without checking if that phrase would match a special
line.  It's okay, becuase it's only used in places where it's already known
that the phrase wont be a special line.  This new rule is not to be used,
except in situations where the text cannot start as a special line.

Summary of changes:
* Updated the treetop grammar to add unsafe_text and use it for horizontal rules
* Fixed to_html, to_s to actually print the unsafe_text after a horizontal rule
* Updated formatting tests to work (added p container)
* Also removed some trailing whitespace from headings tests.  It angers me.